refactor(appkit-ui): move useAiSearchQuery to the beta React entry
The hook's contract is defined entirely by the aiSearch plugin, which ships at
beta from @databricks/appkit/beta. Export the hook and its types from
@databricks/appkit-ui/react/beta so its stability tracks the plugin, instead of
the stable @databricks/appkit-ui/react surface. Consumers and docs updated.
