+++ apps/dev-playground/client/src/routes/ai-search.route.tsx
@@ -5,8 +5,8 @@ import {
   CardHeader,
   CardTitle,
   Input,
-  useAiSearchQuery,
 } from "@databricks/appkit-ui/react";
+import { useAiSearchQuery } from "@databricks/appkit-ui/react/beta";
 import { createFileRoute } from "@tanstack/react-router";
 import { Search } from "lucide-react";
 import { useState } from "react";
