+++ packages/appkit-ui/src/react/hooks/index.ts
@@ -6,12 +6,6 @@ export {
   useResourceStatusToaster,
 } from "../resource-status-indicator";
 export type {
-  AiSearchClientConfig,
-  AiSearchIndexSummary,
-  AiSearchQueryType,
-  AiSearchRequest,
-  AiSearchResponse,
-  AiSearchResult,
   AnalyticsFormat,
   InferResultByFormat,
   InferRowType,
@@ -34,11 +28,6 @@ export {
   type UseAgentChatResult,
   useAgentChat,
 } from "./use-agent-chat";
-export {
-  type UseAiSearchQueryOptions,
-  type UseAiSearchQueryResult,
-  useAiSearchQuery,
-} from "./use-ai-search-query";
 export { useAnalyticsQuery } from "./use-analytics-query";
 export {
   type UseChartDataOptions,
